debug：用户登录后，提交订单orderItem为null

+++ tinyjd-cart-service/src/main/java/com/dcxiaolou/tinyJD/cart/service/impl/CartServiceImpl.java
@@ -92,9 +92,11 @@ public class CartServiceImpl implements CartService {
         omsCartItem.setProductSkuId(productSkuId);
         omsCartItem = cartMapper.selectOne(omsCartItem);
         System.out.println("delCartItemBySkuId: " + omsCartItem);
-        String memberId = omsCartItem.getMemberId();
-        cartMapper.delete(omsCartItem);
-        // 同步缓存
-        flushCartCache(memberId);
+        if (omsCartItem != null) {
+            String memberId = omsCartItem.getMemberId();
+            cartMapper.delete(omsCartItem);
+            // 同步缓存
+            flushCartCache(memberId);
+        }
     }
 }
